Move all /forms to /dataDisplay
